+++ src/index.js
@@ -57,6 +57,8 @@ import CircleAnimate from './Canvas/CircleAnimate';
 import PillarList from './Canvas/PillarList';
 import ScratchOff from './Canvas/ScratchOff';
 import RadarSpan from './Canvas/RadarSpan';
+import HotWords from './Canvas/HeatList';
+import DataScatter from './Canvas/DataScatter';
 
 // import AbstractCanvasVisualization from './Canvas/AbstractCanvasVisualization';
 // import LineCircle from './Canvas/LineCircle';
@@ -114,4 +116,6 @@ module.exports = {
   ThreeModel,
   GlobePointLine,
   RadarSpan,
+  HotWords,
+  DataScatter,
 };