+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 
+
+modify the map rendering scale in ScatterCurveMap
+
 ## 1.0.4
 
 resolve AMapCluster window.AMap.MarkerClusterer is not a constructor