+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ##
+
+add Basic component WorldCloud
+
+Add Pictogram component ProgressBar
+
 add canvas component RadarSpan
 
 add privilege when y axis is [0] in BarChart3D