+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ##
 
+add Basic component WorldCloud
+
+Add Pictogram component ProgressBar
+
+add canvas component RadarSpan
+
 add privilege when y axis is [0] in BarChart3D
 
 modify the map rendering scale in ScatterCurveMap