+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+##
+
+fix isLegendShow bug, add showTooltip and showSymbol two attribute to LineChart
+
+add Basic component WorldCloud
+
+Add Pictogram component ProgressBar
+
+add canvas component RadarSpan
+
+add privilege when y axis is [0] in BarChart3D
+
+modify the map rendering scale in ScatterCurveMap
+
 ## 1.0.4
 
 resolve AMapCluster window.AMap.MarkerClusterer is not a constructor